Add Session Controller
